requirements.txt
google-cloud-vision==3.4.4
pandas==2.0.3
openpyxl==3.1.2
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1